requirements.txt
Django
djangorestframework
Faker
